feat: add GuideDraggablePopover
