+++ apps/builder/src/components/Guide/GuidePopover/index.tsx
@@ -1,7 +1,7 @@
 import { FC, HTMLAttributes } from "react"
 import { useTranslation } from "react-i18next"
 import { useDispatch } from "react-redux"
-import { Button, useMessage, useModal } from "@illa-design/react"
+import { Button, useModal } from "@illa-design/react"
 import {
   actionStyle,
   bottomStyle,
@@ -26,19 +26,19 @@ export const GuidePopover: FC<GuidePopoverProps> = (props) => {
   const modal = useModal()
 
   const handleExitGuide = () => {
-    // modal.show({
-    //   id: "exitGuide",
-    //   title: t("tutorial.modal.tutorial.exit_confirm.title"),
-    //   children: t("tutorial.modal.tutorial.exit_confirm.description"),
-    //   cancelText: t("tutorial.modal.tutorial.exit_confirm.cancel"),
-    //   okText: t("tutorial.modal.tutorial.exit_confirm.exit"),
-    //   okButtonProps: {
-    //     colorScheme: "techPurple",
-    //   },
-    //   onOk: () => {
-    //     dispatch(guideActions.updateGuideStatusReducer(false))
-    //   },
-    // })
+    modal.show({
+      id: "exitGuide",
+      title: t("tutorial.modal.tutorial.exit_confirm.title"),
+      children: t("tutorial.modal.tutorial.exit_confirm.description"),
+      cancelText: t("tutorial.modal.tutorial.exit_confirm.cancel"),
+      okText: t("tutorial.modal.tutorial.exit_confirm.exit"),
+      okButtonProps: {
+        colorScheme: "techPurple",
+      },
+      onOk: () => {
+        dispatch(guideActions.updateGuideStatusReducer(false))
+      },
+    })
   }
 
   return (
