+++ apps/builder/src/components/Guide/WidgetStepMask/index.tsx
@@ -1,0 +1,76 @@
+import { FC, useMemo } from "react"
+import { useTranslation } from "react-i18next"
+import { useSelector } from "react-redux"
+import {
+  applyStepMaskStyle,
+  applyStepMaskWrapperStyle,
+  applyVisibleStyle,
+  ellipsisStyle,
+  moveIconStyle,
+  stepFirstLineStyle,
+} from "@/components/Guide/WidgetStepMask/style"
+import { ReactComponent as MoveIcon } from "@/components/Guide/assets/move.svg"
+import { GUIDE_SELECT_WIDGET } from "@/config/guide/config"
+import { UNIT_HEIGHT } from "@/page/App/components/DotPanel/renderComponentCanvas"
+import { getCanvasShape } from "@/redux/config/configSelector"
+import { BASIC_BLOCK_COLUMNS } from "@/utils/generators/generatePageOrSectionConfig"
+import { WidgetConfig } from "@/widgetLibrary/widgetBuilder"
+
+export interface StepMaskProps {
+  currentStep: number
+}
+
+export const WidgetStepMask: FC<StepMaskProps> = (props) => {
+  const { t } = useTranslation()
+
+  const { currentStep } = props
+  const canvasShape = useSelector(getCanvasShape)
+
+  const unitWidth = useMemo(() => {
+    return canvasShape.canvasWidth / BASIC_BLOCK_COLUMNS
+  }, [canvasShape.canvasWidth])
+
+  const widgetShape = useMemo(() => {
+    return GUIDE_SELECT_WIDGET.map((widget) => {
+      const { w, h } = WidgetConfig[widget].config
+      return {
+        width: w * unitWidth,
+        height: h * UNIT_HEIGHT,
+      }
+    })
+  }, [unitWidth])
+
+  const applyMaskStyle = (step: number) => {
+    return [
+      applyStepMaskStyle(widgetShape[step]),
+      applyVisibleStyle(currentStep <= step),
+    ]
+  }
+
+  return (
+    <div css={applyStepMaskWrapperStyle(currentStep <= 2, unitWidth)}>
+      <div css={stepFirstLineStyle}>
+        <div css={applyMaskStyle(0)}>
+          <MoveIcon css={moveIconStyle} />
+          <span css={ellipsisStyle}>
+            {t("editor.tutorial.panel.onboarding_app.drag_input")}
+          </span>
+        </div>
+        <div css={applyMaskStyle(1)}>
+          <MoveIcon css={moveIconStyle} />
+          <span css={ellipsisStyle}>
+            {t("editor.tutorial.panel.onboarding_app.drag_button")}
+          </span>
+        </div>
+      </div>
+      <div css={applyMaskStyle(2)}>
+        <MoveIcon css={moveIconStyle} />
+        <span css={ellipsisStyle}>
+          {t("editor.tutorial.panel.onboarding_app.drag_table")}
+        </span>
+      </div>
+    </div>
+  )
+}
+
+WidgetStepMask.displayName = "WidgetStepMask"
